pyproject.toml
[tool.poetry]
name = "habits-drf"
version = "0.1.0"
description = ""
authors = ["alexandr <dmitrymaslov2016@mai.ru>"]
readme = "README.md"
packages = [{include = "habits_drf"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.6"
djangorestframework-simplejwt = "^5.2.2"
django-cors-headers = "^4.2.0"
drf-yasg = "^1.21.7"
requests = "^2.31.0"
celery = "^5.3.1"
redis = "^4.6.0"
django-celery-beat = "^2.5.0"
flake8-for-pycharm = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
